# api-contract/openapi.json
{
  "components" : {
    "schemas" : {
...
      "AdminDepositRequestResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "cancelled_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "confirmation_reference" : {
            "type" : "string"
          },
          "confirmed_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "confirmed_by_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "deposit_method_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deposit_method_title" : {
            "type" : "string"
          },
          "details_issued_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "details_issued_by_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "events" : {
            "items" : {
              "$ref" : "#/components/schemas/MoneyOperationEventResponse"
            },
            "type" : "array"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "payment_details" : {
            "type" : "string"
          },
          "payment_instruction" : {
            "$ref" : "#/components/schemas/DepositPaymentInstructionResponse"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reject_reason" : {
            "type" : "string"
          },
          "rejected_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "rejected_by_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PENDING_DETAILS", "WAITING_PAYMENT", "PAYMENT_VERIFICATION", "CONFIRMED", "REJECTED", "EXPIRED", "CANCELLED" ],
            "type" : "string"
          },
          "treasury_account_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "treasury_transaction_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "treasury_transactions" : {
            "items" : {
              "$ref" : "#/components/schemas/TreasuryTransactionResponse"
            },
            "type" : "array"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "user_marked_paid_at" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "AdminDepositRequestShortResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "cancelled_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "confirmed_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "deposit_method_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deposit_method_title" : {
            "type" : "string"
          },
          "details_issued_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "rejected_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING_DETAILS", "WAITING_PAYMENT", "PAYMENT_VERIFICATION", "CONFIRMED", "REJECTED", "EXPIRED", "CANCELLED" ],
            "type" : "string"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserSummary"
          },
          "user_marked_paid_at" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "BackofficeWithdrawalRequestResponse" : {
        "properties" : {
          "actual_payout_amount" : {
            "type" : "number"
          },
          "amount" : {
            "type" : "number"
          },
          "cancelled_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "completed_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "events" : {
            "items" : {
              "$ref" : "#/components/schemas/MoneyOperationEventResponse"
            },
            "type" : "array"
          },
          "method_note" : {
            "type" : "string"
          },
          "opened_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "payout_plan" : {
            "$ref" : "#/components/schemas/WithdrawalPayoutPlanResponse"
          },
          "payout_profile_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "processed_by_user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "processing_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "rejected_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "rejection_reason" : {
            "type" : "string"
          },
          "requisites" : {
            "additionalProperties" : { },
            "type" : "object"
          },
          "status" : {
            "type" : "string"
          },
          "treasury_account_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "treasury_transaction_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "treasury_transactions" : {
            "items" : {
              "$ref" : "#/components/schemas/TreasuryTransactionResponse"
            },
            "type" : "array"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "user_account_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "user_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "withdrawal_method_code" : {
            "type" : "string"
          },
          "withdrawal_method_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "withdrawal_method_title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ConfirmDepositRequest" : {
        "properties" : {
          "confirmation_reference" : {
            "type" : "string"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "treasury_amount" : {
            "minimum" : 1.0E-4,
            "type" : "number"
          },
          "treasury_external_reference" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ConfirmWithdrawalRequest" : {
        "properties" : {
          "actual_payout_amount" : {
            "minimum" : 1.0E-4,
            "type" : "number"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "treasury_amount" : {
            "minimum" : 1.0E-4,
            "type" : "number"
          },
          "treasury_external_reference" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "DepositRequestResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "cancelled_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "confirmed_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "created_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency_code" : {
            "type" : "string"
          },
          "deposit_method_id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "deposit_method_title" : {
            "type" : "string"
          },
          "details_issued_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "payment_details" : {
            "type" : "string"
          },
          "payment_instruction" : {
            "$ref" : "#/components/schemas/DepositPaymentInstructionResponse"
          },
          "public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reject_reason" : {
            "type" : "string"
          },
          "rejected_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING_DETAILS", "WAITING_PAYMENT", "PAYMENT_VERIFICATION", "CONFIRMED", "REJECTED", "EXPIRED", "CANCELLED" ],
            "type" : "string"
          },
          "updated_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "user_marked_paid_at" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "IssueDetailsRequest" : {
        "properties" : {
          "deposit_payment_route_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "expires_at" : {
            "format" : "date-time",
            "type" : "string"
          },
          "operator_comment" : {
            "type" : "string"
          },
          "payment_details" : {
            "type" : "string"
          },
          "treasury_account_public_id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "treasury_amount" : {
            "minimum" : 1.0E-4,
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "PageAdminDepositRequestShortResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/AdminDepositRequestShortResponse"
            },
            "type" : "array"
          },
          "empty" : {
            "type" : "boolean"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "numberOfElements" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PageableObject" : {
        "properties" : {
          "offset" : {
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "pageSize" : {
            "format" : "int32",
            "type" : "integer"
          },
          "paged" : {
            "type" : "boolean"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "RejectDepositRequest" : {
        "properties" : {
          "operator_comment" : {
            "type" : "string"
          },
          "reject_reason" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "reject_reason" ],
        "type" : "object"
      },
...
      "SortObject" : {
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "UserSummary" : {
        "properties" : {
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "username" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Generated API contract for frontend/backend synchronization.",
    "title" : "Prime Market Backend API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/admin/deposit-requests" : {
      "get" : {
        "operationId" : "list_7",
        "parameters" : [ {
          "description" : "Canonical multi-value status filter. Supports repeated `status` params and comma-separated values in a single parameter.",
          "example" : "PENDING_DETAILS,WAITING_PAYMENT",
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "items" : {
              "enum" : [ "PENDING_DETAILS", "WAITING_PAYMENT", "PAYMENT_VERIFICATION", "CONFIRMED", "REJECTED", "EXPIRED", "CANCELLED" ],
              "type" : "string"
            },
            "type" : "array"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "createdAt,DESC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageAdminDepositRequestShortResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
    "/api/admin/deposit-requests/{publicId}" : {
      "get" : {
        "operationId" : "get_4",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
    "/api/admin/deposit-requests/{publicId}/confirm" : {
      "post" : {
        "operationId" : "confirm_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmDepositRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
    "/api/admin/deposit-requests/{publicId}/issue-details" : {
      "post" : {
        "operationId" : "issueDetails",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IssueDetailsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
    "/api/admin/deposit-requests/{publicId}/reject" : {
      "post" : {
        "operationId" : "reject_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RejectDepositRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
...
    "/api/backoffice/deposit-requests" : {
      "get" : {
        "operationId" : "list_6",
        "parameters" : [ {
          "description" : "Canonical multi-value status filter. Supports repeated `status` params and comma-separated values in a single parameter.",
          "example" : "PENDING_DETAILS,WAITING_PAYMENT",
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "items" : {
              "enum" : [ "PENDING_DETAILS", "WAITING_PAYMENT", "PAYMENT_VERIFICATION", "CONFIRMED", "REJECTED", "EXPIRED", "CANCELLED" ],
              "type" : "string"
            },
            "type" : "array"
          }
        }, {
          "description" : "Zero-based page index (0..N)",
          "in" : "query",
          "name" : "page",
          "required" : false,
          "schema" : {
            "default" : 0,
            "minimum" : 0,
            "type" : "integer"
          }
        }, {
          "description" : "The size of the page to be returned",
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "default" : 20,
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "in" : "query",
          "name" : "sort",
          "required" : false,
          "schema" : {
            "default" : [ "createdAt,DESC" ],
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageAdminDepositRequestShortResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
    "/api/backoffice/deposit-requests/{publicId}" : {
      "get" : {
        "operationId" : "get_3",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
    "/api/backoffice/deposit-requests/{publicId}/confirm" : {
      "post" : {
        "operationId" : "confirm_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmDepositRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
    "/api/backoffice/deposit-requests/{publicId}/issue-details" : {
      "post" : {
        "operationId" : "issueDetails_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/IssueDetailsRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
    "/api/backoffice/deposit-requests/{publicId}/reject" : {
      "post" : {
        "operationId" : "reject_2",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RejectDepositRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminDepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-admin-controller" ]
      }
    },
...
    "/api/backoffice/withdrawal-requests/{publicId}/confirm" : {
      "post" : {
        "operationId" : "confirm",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConfirmWithdrawalRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BackofficeWithdrawalRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "backoffice-withdrawal-request-controller" ]
      }
    },
...
    "/api/deposit-requests/{publicId}" : {
      "get" : {
        "operationId" : "get_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "publicId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DepositRequestResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "deposit-request-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Relative to the current backend base URL",
    "url" : "/"
  } ]
}